Fix return dist logging

+++ rad_embeddings/ppo.py
@@ -323,7 +323,7 @@ def make_train(config, env, network, batchify):
                     timesteps = info["timestep"][-1, :]
                     log["timestep"] = int(np.sum(timesteps) / config["NUM_AGENTS"])
 
-                    returns = [int(x) for x in return_buffer_debug]
+                    returns = [float(x) for x in return_buffer_debug]
                     n = len(returns)
                     counts = Counter(returns)
                     return_dist = {i: float(counts[i])/float(n) for i in counts}
